+++ app/src/main/java/com/example/mymovieapp/movielist/ui/MovieFragment.kt
@@ -8,7 +8,6 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.mymovieapp.movielist.data.models.MovieAdapter
 import com.example.mymovieapp.databinding.MovieFragmentBinding
 import com.example.mymovieapp.movielist.data.models.Movie
 import com.example.mymovieapp.utils.Status
